What I could salvage from my project
Due to upgrading my PC, I did not consider keeping my data from my old PC to have a new fresh start. This meant that I lost the original files of my prototype, but fortunately I was still able to download the application since I sent it to myself for a showcase on a different laptop.